refactor: convert palindrome_checker function into PalindromeChecker class
For the moment, I've merely indented the palindrome_checker function
within the class definition and added self as another parameter.

I'm motivated to refactor the project out of a desire to align my code
with the single responsibility principle for functions. A lot of the
SRP functions I can extract from the function as it stands are helper
functions, and could do with being private methods in a class.
I'm also convicted it will make testing the code easier in the long run.